+++ lib/src/orm/query/query_builder.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/foundation.dart';
-import '../../database/connection.dart';
 
 class QueryBuilder {
-  Connection connection;
-
   Map<String, List> _bindings = {
     'select': [],
     'from': [],
@@ -47,33 +44,28 @@ class QueryBuilder {
     '!~~*',
   ];
 
-  List<String> columns;
-  bool distinct;
-  String table;
+  final List<String> _columns = [];
+  String _table;
+  bool _distinct = false;
   final List joins = [];
   final List<Map<String, String>> _whereColumns = [];
-  final List<dynamic> wheresArgs = [];
-  List groups;
-  List havings;
-  int limit;
-  int offset;
+  final List<dynamic> _wheresArgs = [];
+  final List _groups = [];
+  final List<Map<String, String>> _orders = [];
+  final List<Map<String, dynamic>> _havings = [];
+  int _limit;
+  int _offset;
   List unions;
   int unionLimit;
   int unionOffset;
-
-  QueryBuilder({@required this.connection});
-
-  QueryBuilder select({List<String> columns = const ['*']}) {
-    this.columns = columns;
-    return this;
-  }
 
   QueryBuilder from({String table}) {
     table = table;
     return this;
   }
 
-  QueryBuilder selectRaw({String expression}) {
+  QueryBuilder select(List<String> columns) {
+    this._columns.addAll(columns);
     return this;
   }
 
@@ -90,9 +82,114 @@ class QueryBuilder {
       'argsSize': values.length.toString()
     });
 
-    values.forEach((v) => wheresArgs.add(v));
+    values.forEach((v) => _wheresArgs.add(v));
 
     return this;
+  }
+
+  QueryBuilder orderBy({String column, String type, List<String> columns}) {
+    assert((column != null && columns == null) ||
+            (column == null && columns != null)
+        ? true
+        : throw 'QueryBuilder: set column or columns field');
+    if (columns == null)
+      _orders.add({'column': column, 'type': type});
+    else
+      columns.forEach((col) {
+        _orders.add({'column': col, 'type': type});
+      });
+    return this;
+  }
+
+  QueryBuilder groupBy(List<String> columns) {
+    _groups.addAll(columns);
+    return this;
+  }
+
+  QueryBuilder having(
+      {@required String column,
+      String operator = '=',
+      @required dynamic value,
+      String condition = 'AND'}) {
+    _havings.add({
+      'column': column,
+      'operator': operator,
+      'condition': condition,
+      'value': value
+    });
+
+    return this;
+  }
+
+  List<String> get columns {
+    final List clms = _columns.toList();
+    _columns.clear();
+    return clms;
+  }
+
+  set setDistinct(bool val) => _distinct = val;
+  bool get distinct {
+    final bool dst = _distinct;
+    _distinct = false;
+    return dst;
+  }
+
+  set setLimit(int val) => _limit = val;
+  int get limit {
+    final int lmt = _limit;
+    _limit = null;
+    return lmt;
+  }
+
+  set setOffset(int val) => _offset = val;
+  int get offset {
+    final int offt = _offset;
+    _offset = null;
+    return offt;
+  }
+
+  String get groups {
+    final List grps = _groups.toList();
+    _groups.clear();
+    return grps.join(',');
+  }
+
+  String get orders {
+    List<String> ords = _orders.map((grp) {
+      String type = grp['type'] == null ? '' : grp['type'].toLowerCase();
+      type = type != 'asc' && type != 'desc' ? '' : type;
+      return grp['column'] + (type.length > 0 ? ' $type' : '');
+    }).toList();
+    _orders.clear();
+    return ords.join(',');
+  }
+
+  String get havings {
+    StringBuffer str = StringBuffer();
+    int size = _havings.length;
+    int cnt = 0;
+
+    _havings.forEach((grp) {
+      str.write('${grp['column']}');
+      str.write(' ');
+      str.write(grp['operator']);
+      str.write(' ');
+
+      if (grp['value'].runtimeType == String)
+        str.write("'${grp['value']}'");
+      else
+        str.write(grp['value']);
+
+      if (size > 1) {
+        String cond = _havings[++cnt]['condition'];
+        str.write(' ');
+        str.write(cond);
+        str.write(' ');
+        --size;
+      }
+    });
+    _havings.clear();
+    return str.toString();
   }
 
   /// Prepare a String for the where condition
@@ -123,9 +220,14 @@ class QueryBuilder {
         --size;
       }
     });
-
-    print(str);
+    _whereColumns.clear();
     return str.toString();
+  }
+
+  List<dynamic> get wheresArgs {
+    final List<dynamic> args = _wheresArgs.toList();
+    _wheresArgs.clear();
+    return args;
   }
 
   /// Generate where IN ---> IN (?,?,..)