+++ CHANGELOG.md
@@ -1,5 +1,14 @@
+## [0.0.3] - 2020-05-2.
+
+* **Created the following model methods:**
+* select, count, min, avg, max
+* where, orWhere, whereIn, orWhereIn, whereNotIn, orWhereNotIn
+* insert, update, get
+* distinct, groupBy, orderBy, orderByDesc, orderByAsc
+* having, orHaving
+
 ## [0.0.2] - 2020-04-18.
-* Created ConnectHelper Class. Possibility to manage the connection with SQflite.
+* Created Connection Class. Possibility to manage the connection with SQflite.
 
 ## [0.0.1] - 2020-04-9.
 * Created Database Manager.